requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
executing==2.2.1
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.5
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pywin32==311
pyzmq==27.0.2
setuptools==80.10.2
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
wcwidth==0.2.13
wheel==0.46.3

# Web framework
fastapi
uvicorn[standard]
python-multipart # for file upload endpoints

# Config
python-dotenv

# Database
sqlalchemy

# LangChain ecosystem
langchain
langchain-core
langchain-openai
langchain-community
langgraph

# Vector DB
qdrant-client

# Hybrid retrieval
rank_bm25                 # required by LangChain's BM25Retriever
sentence-transformers     # required by HuggingFaceCrossEncoder reranker

# Web search fallback (CRAG Tavily)
tavily-python

# Document parsing
beautifulsoup4
requests
pypdf
python-docx

# Evaluation
ragas
datasets

# Testing
pytest

# Docling
docling